[feat/#12] comment 내용을 projects app으로 이동
